requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysqlclient
python-multipart
python-jose
passlib
bcrypt